completed endpoint for editing venue, moved body validation checkers to utils/validation
