+++ backend/utils/validation.js
@@ -1,5 +1,6 @@
 // backend/utils/validation.js
 const { validationResult } = require('express-validator');
+const { check } = require('express-validator'); //testing
 
 // middleware for formatting errors from express-validator middleware
 // (to customize, see express-validator's documentation)
@@ -26,6 +27,56 @@ const handleValidationErrors = (req, _res, next) => {
   next();
 };
 
-module.exports = {
+/************************* BODY VALIDATORS ******************* */
+const validateGroup = [
+  check('name')
+    .exists({ checkFalsy: true })
+    .isLength({max: 60, min: 1}) //inclusive
+    .withMessage("Name must be 60 characters or less"),
+  check('about')
+    .exists({ checkFalsy: true })
+    .isLength({min: 50}) //inclusive
+    .withMessage("About must be 50 characters or more"),
+  check('type')
+    .exists({ checkFalsy: true })
+    .isIn(['Online', 'In person'])
+    .withMessage("Type must be 'Online' or 'In person'"),
+  check('private')
+    .exists({ checkFalsy: false })
+    .isBoolean() //if want strictly true, false, use {strict: true}
+    .withMessage("Private must be a boolean"),
+  check('city')
+    .exists({ checkFalsy: true })
+    .withMessage("City is required"),
+  check('state')
+    .exists({ checkFalsy: true })
+    .withMessage("State is required"),
   handleValidationErrors
+];
+
+const validateVenue = [
+  check('address')
+    .exists({ checkFalsy: true })
+    .withMessage("Street address is required"),
+  check('city')
+    .exists({ checkFalsy: true })
+    .withMessage("City is required"),
+  check('state')
+    .exists({ checkFalsy: true })
+    .withMessage("State is required"),
+  check('lat')
+    .exists({ checkFalsy: false })
+    .isFloat({min:-90, max:90})
+    .withMessage("Latitude is not valid"),
+  check('lng')
+    .exists({ checkFalsy: false })
+    .isFloat({min:-180, max:180})
+    .withMessage("Longitude is not valid"),
+  handleValidationErrors
+];
+
+module.exports = {
+  handleValidationErrors,
+  validateGroup,
+  validateVenue,
 };
